+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.11]
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+
 ### Added
 
 ### Changed
@@ -227,7 +242,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/nelknet/Nelknet.LibSQL/compare/v0.2.10...HEAD
+[Unreleased]: https://github.com/nelknet/Nelknet.LibSQL/compare/v0.2.11...HEAD
+[0.2.11]: https://github.com/nelknet/Nelknet.LibSQL/compare/v0.2.10...v0.2.11
 [0.2.10]: https://github.com/nelknet/Nelknet.LibSQL/compare/v0.2.9...v0.2.10
 [0.2.9]: https://github.com/nelknet/Nelknet.LibSQL/compare/v0.2.8...v0.2.9
 [0.2.8]: https://github.com/nelknet/Nelknet.LibSQL/compare/v0.2.7...v0.2.8